test_requirements.txt
-e .
pytest-django
factory_boy
sphinx
